client/queue/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
urllib3==2.0.7
PyYAML==6.0.1
pykafka==2.8.0
